requirements.txt
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
supabase>=2.6.0
psycopg-binary>=3.2.0
scikit-learn>=1.3.0
xgboost>=2.0.0

#Authentication için
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.5.0